userTypeId Fix 2

+++ ApplicationChallenge/Models/Permissions.cs
@@ -10,7 +10,6 @@ namespace ApplicationChallenge.Models
     {
         public int Id { get; set; }
         public string Title { get; set; }
-
-        public int UserTypeId { get; set; }
+        public long UserTypeId { get; set; }
     }
 }
